Macro Function and Formatting/Calculations.
I once again changed the calculations because I found a newer, more accurate standard, the Mifflin-St Jeor Equation. Along with this I added an extra activity level as this is what it is based off of.  I also started my macronutrient function that will determine your macros based on your calories and activity level. To be able to pass the information to this function that I needed, I had to change a few variables around. I also had to completely change the call function at the end. due to other functions now returning multiple values.