User Error Proofing
While testing my program to make sure everything was running smoothly, I noticed anytime I made a mistake inputting information, I would have to restart the program. To counter act this I put while loops for each input message. This way it iterates through each input message until a valid input is submitted before moving on to the next while loop to do it again.  To make sure each input is valid, I check to make sure it fits the needed criteria of that information. For example for age it checks that its a number and that it is greater than 0. This is due to someone being that young could not be using this app and it would mess up calculations if a user accidentally inputted a 0 instead of say, 20 and they didn't realize.