Added a choice to use imperial or metric system measurments.
I had previously just taken in metric measurements and converted them to imperial for ease with the formulas. To make the program more user friendly and diverse, the user now determines which measurement system to use to make it easier for the user. I did this by just adding a few if else blocks along with while loops to make sure their inputs are valid, The if and elif determines whether or not the user inputted imperial or metric. Then the program asks you the prompts that are unit dependent such as weight and height in the units that correlate with the inputted measuring system,