Added ability for users to save their profile.
Using the skills we learned earlier in the semester, working with files in python. I added a load_profile and a save_profile function. The save_profile function takes in six parameters that are calculated earlier in the program and saves the information to a file that is names after the user's first name. The load profile function checks when the user inputs their name to see if they already have a file and if so it loads up the contents of that file in the terminal. If the user does not already have a file saved, it will then go through the program asking the user all the prompts with the ability to save the information in a file at the end.