Added GUI for name input
Using tkinter I added a GUI for the part of the system where it asks for the name to see if the user already has a profile. If not it will give them a warning window so they know to go back to the program and fill it out. If the user does already have a file it will then load the file in the GUI.