Added GUI for all prompts and information
I didn't mean to commit the python file in my last commit. It was supposed to just be the .txt file. So this is my description on what was done with the last commit.  At first I had to solve the issue of user no longer being defined in line 202. I did so by returning user from the function I had put it in and then assigned user the value that the function returns in line 196. To build a GUI for all the inputs I used two tkinter modules that I imported, messagebox and simpledialog. I use these functions to replace my input and print commands to pop up a GUI to ask the prompts and show information. A roadblock I ran into was that the functions from the module only took 2 parameters and they way I had formatted my input and print statements made it seem as there was more arguments because I used commas. To fix this I had to concatenate strings and because some of my variables were integers, I had to cast them to strings before they could be concatenated.